requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
pyyaml>=6.0.1
tqdm>=4.66.0
urllib3>=2.1.0
certifi>=2023.11.17